CTFs/Ethernaut/Coin_Flip_DONE/pyproject.toml
[tool.poetry]
name = "coin-flip"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "coin_flip"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
eth-ape = "^0.6.10"
ape-alchemy = "^0.6.1"
ape-ganache = "^0.6.6"
ape-solidity = "^0.6.5"
ape-hardhat = "^0.6.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
